confirm before delete snippet

+++ assets/js/profile.js
@@ -124,27 +124,31 @@ $(function() {
 
     _mineSnippetsDiv.on('click', 'a.link-delete-snippet', function() {
         var self = $(this);
-        var snippetId = self.attr('data-snippet-id');
-        $.ajax({
-            type: 'DELETE',
-            url: '/api/snippets',
-            data: {
-                _csrf: _token.val(),
-                snippetId: snippetId
-            },
-            dataType: 'json',
-            success: function(data) {
-                if (data.code === 200) {
-                    self.parent().parent().remove();
-                    bootbox.alert(Message.DELETE_SNIPPET_SUCCESS);
-                } else if (data.code === 400) {
-                    bootbox.alert(Message.SNIPPET_NOT_EXSIT);
-                } else {
-                    bootbox.alert(Message.DELETE_SNIPPET_FORBIDDEN);
-                }
-            },
-            error: function(xhr, status, err) {
-                bootbox.alert(Message.SERVER_ERROR);
+        bootbox.confirm(Message.DELETE_SNIPPET_CONFIRM, function(result) {
+            if (result) {
+                var snippetId = self.attr('data-snippet-id');
+                $.ajax({
+                    type: 'DELETE',
+                    url: '/api/snippets',
+                    data: {
+                        _csrf: _token.val(),
+                        snippetId: snippetId
+                    },
+                    dataType: 'json',
+                    success: function(data) {
+                        if (data.code === 200) {
+                            self.parent().parent().remove();
+                            bootbox.alert(Message.DELETE_SNIPPET_SUCCESS);
+                        } else if (data.code === 400) {
+                            bootbox.alert(Message.SNIPPET_NOT_EXSIT);
+                        } else {
+                            bootbox.alert(Message.DELETE_SNIPPET_FORBIDDEN);
+                        }
+                    },
+                    error: function(xhr, status, err) {
+                        bootbox.alert(Message.SERVER_ERROR);
+                    }
+                });
             }
         });
     });
